Fix getZvalAsDouble truncating fractional floats to integers
The helper was declared to return 'long' despite reading Z_DVAL into
a local 'double'. C truncated the double to long on return; the
truncation was invisible because every caller assigned the result
back to a double target — so 1.5 became 1, 2.7 became 2, and so on.

Affected call sites:
  - GEOSCoordSeq::copyFromArrays (field-reported regression — fractional
    coordinates were silently rounded to integers).
  - GEOSGeometry::buffer 'mitre_limit' style key (fractional limits were
    rounded, e.g. 1.5 became 1.0).
  - GEOSGeometry::coverageClean 'snap_distance' and 'gap_max_width' params.

Fix is a one-line return-type correction. All callers already expected
a double — no further changes needed.

Regression test added:
  tests/006_CoordSeq.phpt::testCopyFromArraysPreservesFractionalFloats.

25/25 .phpt tests pass.

+++ geos.c
@@ -238,7 +238,7 @@ static long getZvalAsLong(GEOS_PHP_ZVAL val)
     return ret;
 }
 
-static long getZvalAsDouble(GEOS_PHP_ZVAL val)
+static double getZvalAsDouble(GEOS_PHP_ZVAL val)
 {
     double ret;
     zval tmp;
